Rename p11ez -> p11
